Fix enum cases to uppercase with migrations
