+++ api/graphql/types/reservations/serializers/staff_reservation_modify_serializers.py
@@ -7,11 +7,15 @@ from api.graphql.extensions.duration_field import DurationField
 from api.graphql.extensions.legacy_helpers import OldChoiceCharField, OldPrimaryKeyUpdateSerializer
 from api.graphql.extensions.validation_errors import ValidationErrorCodes, ValidationErrorWithCode
 from api.graphql.types.reservations.serializers.mixins import ReservationSchedulingMixin
-from applications.choices import CustomerTypeChoice
 from applications.models import City
 from common.fields.serializer import IntegerPrimaryKeyField
 from reservation_units.models import ReservationUnit
-from reservations.choices import RESERVEE_LANGUAGE_CHOICES, ReservationStateChoice, ReservationTypeChoice
+from reservations.choices import (
+    RESERVEE_LANGUAGE_CHOICES,
+    CustomerTypeChoice,
+    ReservationStateChoice,
+    ReservationTypeChoice,
+)
 from reservations.models import AgeGroup, Reservation, ReservationPurpose
 
 DEFAULT_TIMEZONE = get_default_timezone()
